feat(lunox-core): separate core module, facades, and contracts
