+++ packages/lunox-core/src/Contracts/index.ts
@@ -11,3 +11,4 @@ export * from "./Encryption";
 export * from "./Request";
 export * from "./Validation";
 export * from "./Response";
+export * from "../Types";
