--- packages/lunox-core/src/Support/Facades/Validator.ts
@@ -1,11 +1,0 @@
-import Factory from "../../Validation/Factory";
-import Facade from "./Facade";
-import useFacade from "./useFacade";
-
-class Validator extends Facade {
-  public static getFacadeAccessor() {
-    return Factory.symbol;
-  }
-}
-
-export default useFacade<Factory>(Validator);
