+++ packages/lunox-core/src/Support/Facades/index.ts
@@ -4,7 +4,6 @@ import Response from "./Response";
 import Route from "./Route";
 import Storage from "./Storage";
 import useFacade from "./useFacade";
-import Validator from "./Validator";
 import { ExtendedFacade } from "./Facade";
 import Config from "./Config";
 import Als from "./Als";
@@ -16,7 +15,6 @@ export {
   Route,
   Storage,
   useFacade,
-  Validator,
   Crypt,
   Config,
 };
