+++ packages/lunox-core/src/Types.ts
@@ -1,6 +1,5 @@
 import type { Polka } from "polka";
-import type { Request } from "./Http/Request";
-export * from "./Contracts";
+import type { Request } from "./Contracts/Request";
 export type Concrete = NewableFunction | (() => any);
 
 export type CallBack = (...params: any[]) => any;
